+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * removed `v` attribute from manifest
   * ignore non-existing source directories in `localizer` task
   * ignore non-existing library paths in `server` task
+  * the value set for the `snapshotRepoUrl` option is no longer ignored
   * replaced usages of deprecated `groovy` configuration by `compile` configuration
   * added `org.jenkins-ci.jpi` as alternative qualified plugin id for Gradle plugin portal inclusion
   * `jpiDeployUser` and `jpiDeployPassword` properties from `org.jenkinsci.gradle.plugins.jpi.JpiExtension` were not